requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==108.0.5359.71.0
click==8.1.3
colorama==0.4.6
cramjam==2.6.2
decorator==5.1.1
exceptiongroup==1.1.0
fastparquet==2022.12.0
Flask==2.2.2
frozenlist==1.3.3
fsspec==2022.11.0
gcsfs==2022.11.0
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.0
oauthlib==3.2.2
outcome==1.2.0
packaging==22.0
pandas==1.5.2
protobuf==4.21.12
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
wsproto==1.2.0
yarl==1.8.2
